backend/requirements.txt
flask
flask-cors
sentence-transformers
scikit-learn
requests
pypdf
groq
python-dotenv